Move common compose bits into a library...
... so that other components of our stack can make use of it.
