+++ Cargo.toml
@@ -15,5 +15,6 @@ members = [
 	"rpc",
 	"sysfs",
 	"services",
-	"mbus-api"
+	"mbus-api",
+    "composer",
 ]
