+++ composer/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "composer"
+version = "0.1.0"
+authors = ["Tiago Castro <tiago.castro@mayadata.io>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+tokio = { version = "0.2", features = ["full"] }
+futures = "0.3.8"
+tonic = "0.1"
+crossbeam = "0.7.3"
+rpc = { path = "../rpc" }
+ipnetwork = "0.17.0"
+bollard = "0.8.0"
