+++ mayastor/Cargo.toml
@@ -110,3 +110,4 @@ version = "0.7"
 [dev-dependencies]
 assert_matches = "1.2"
 run_script = "*"
+composer = { path = "../composer" }
